tests/wpt/tests/html/semantics/popovers/popover-open-in-beforetoggle.html

<!DOCTYPE html>
<meta charset="utf-8" />
<title>Popover beforetoggle event opening new popovers</title>
<link rel="author" href="mailto:masonf@chromium.org">
<link rel=help href="https://html.spec.whatwg.org/multipage/indices.html#event-beforetoggle">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="resources/popover-utils.js"></script>

<div popover id=p1>Popover 1
  <div popover id=p2>Popover 2
    <div popover id=p3>Popover 3</div>
  </div>
</div>
<div id=outside>Outside</div>
<dialog id=dialog>Dialog</dialog>

<script>
  function getSignal(t) {
    const controller = new AbortController();
    t.add_cleanup(() => controller.abort());
    return controller.signal;
  }

  test((t) => {
    p1.showPopover();
    p1.addEventListener('beforetoggle',() => {
      assert_throws_dom("InvalidStateError",() => p2.showPopover());
    },{signal: getSignal(t)});
    p1.hidePopover(); // Potential crash here
    assert_false(p1.matches(':popover-open'));
    assert_false(p2.matches(':popover-open'));
  },'Open popover from closing beforetoggle event');

  test((t) => {
    p1.showPopover();
    p1.addEventListener('beforetoggle',() => {
      assert_throws_dom("InvalidStateError",() => p2.showPopover());
    },{signal: getSignal(t)});
    p2.addEventListener('beforetoggle',() => {
      assert_throws_dom("InvalidStateError",() => p3.showPopover());
    },{signal: getSignal(t)});
    p1.hidePopover(); // Potential crash here
    assert_false(p1.matches(':popover-open'));
    assert_false(p2.matches(':popover-open'));
    assert_false(p3.matches(':popover-open'));
  },'Open double-nested popovers from closing beforetoggle event');

  test(t => {
    p1.showPopover();
    p1.addEventListener('beforetoggle',() => {
      assert_throws_dom("InvalidStateError",() => p2.showPopover());
    },{signal: getSignal(t)});
    p2.addEventListener('beforetoggle',() => {
      assert_throws_dom("InvalidStateError",() => p3.showPopover());
    },{signal: getSignal(t)});
    dialog.showModal(); // Potential crash here
    assert_false(p1.matches(':popover-open'));
    assert_false(p2.matches(':popover-open'));
    assert_false(p3.matches(':popover-open'));
    dialog.close();
  },'Open double-nested popovers from closing beforetoggle event, dialog open');

  promise_test(async t => {
    p1.showPopover();
    p1.addEventListener('beforetoggle',() => {
      assert_throws_dom("InvalidStateError",() => p2.showPopover());
    },{signal: getSignal(t)});
    p2.addEventListener('beforetoggle',() => {
      assert_throws_dom("InvalidStateError",() => p3.showPopover());
    },{signal: getSignal(t)});
    await clickOn(outside); // Potential crash here
    assert_false(p1.matches(':popover-open'));
    assert_false(p2.matches(':popover-open'));
    assert_false(p3.matches(':popover-open'));
  },'Open double-nested popovers from closing beforetoggle event, light dismiss');
</script>

<div id=p4 popover=auto>p4</div>
<div id=p5 popover=auto>p5</div>
<div id=p6 popover=hint>p6</div>
<script>
test(() => {
  assert_false(p4.matches(':popover-open'));
  assert_false(p5.matches(':popover-open'));
  assert_false(p6.matches(':popover-open'));

  p4.showPopover();
  p4.onbeforetoggle = (e) => {
    if (e.newState === "closed") {
      assert_throws_dom("InvalidStateError",() => p6.showPopover());
    }
  };
  assert_true(p4.matches(':popover-open'), 'p4 should be open');
  p5.showPopover();

  assert_false(p4.matches(':popover-open'), 'p4 should be closed (by opening p5)');
  assert_true(p5.matches(':popover-open'), 'p5 should be open');
  assert_false(p6.matches(':popover-open'), 'p6 should be closed');
  p5.hidePopover(); // Cleanup
}, 'Showing a hint during a showPopover that hides an auto should throw InvalidStateError');
</script>
